Move intergration.py to corresponding modules utils.py
