+++ src/megatron/bridge/training/utils/train_utils.py
@@ -18,12 +18,14 @@ import math
 import os
 import time
 from collections import defaultdict
+from collections.abc import Callable
 from datetime import datetime
 from functools import partial
 from typing import TYPE_CHECKING, Any, Optional, Union
 
 import torch
 import torch.nn as nn
+from megatron.core import tensor_parallel
 from megatron.core.num_microbatches_calculator import get_num_microbatches
 from megatron.core.tensor_parallel import param_is_not_tensor_parallel_duplicate
 from megatron.core.transformer.module import MegatronModule
@@ -46,6 +48,148 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
+
+ModelList = list[MegatronModule]
+ModelHook = Callable[[ModelList], ModelList | None]
+
+
+class LinearForLastLayer(nn.Linear):
+    """Final replicated projection head compatible with Megatron output-layer calls.
+
+    Megatron-Core output layers receive a few runtime-only arguments. This head
+    accepts those arguments for call-site compatibility while using a standard
+    replicated linear projection.
+    """
+
+    def __init__(self, *, input_size: int, output_size: int, sequence_parallel: bool) -> None:
+        """Initialize a replicated final projection.
+
+        Args:
+            input_size: Hidden dimension of the transformer output.
+            output_size: Output dimension of the value/reward head.
+            sequence_parallel: Whether to gather sequence-parallel activations.
+        """
+        super().__init__(in_features=input_size, out_features=output_size, bias=False)
+        self.sequence_parallel = sequence_parallel
+        if sequence_parallel:
+            setattr(self.weight, "sequence_parallel", True)
+
+    def forward(
+        self,
+        input_: torch.Tensor,
+        weight: torch.Tensor | None = None,
+        runtime_gather_output: bool | None = None,
+    ) -> tuple[torch.Tensor, None]:
+        """Run the final projection and return Megatron-style ``(output, bias)``."""
+        del weight, runtime_gather_output
+        logits = super().forward(input_).float()
+        if self.sequence_parallel:
+            logits = tensor_parallel.gather_from_sequence_parallel_region(
+                logits,
+                tensor_parallel_output_grad=False,
+            )
+        return logits, None
+
+
+def create_value_head_hook(*, hidden_size: int, sequence_parallel: bool, output_size: int = 1) -> ModelHook:
+    """Create a pre-wrap hook that replaces the final pipeline stage output head.
+
+    Args:
+        hidden_size: Hidden dimension of the transformer output.
+        sequence_parallel: Whether the model uses sequence parallelism.
+        output_size: Number of outputs produced by the final head.
+
+    Returns:
+        A model hook suitable for external trainer provider construction.
+    """
+    from megatron.core import parallel_state
+
+    _register_linear_for_last_layer_mapping()
+
+    def hook(model: ModelList | MegatronModule) -> ModelList:
+        model_chunks = _ensure_model_list(model)
+        model_post_process: list[bool] = []
+        if (
+            parallel_state.get_pipeline_model_parallel_world_size() > 1
+            and parallel_state.get_virtual_pipeline_model_parallel_world_size() is not None
+        ):
+            for vp_stage in range(parallel_state.get_virtual_pipeline_model_parallel_world_size()):
+                model_post_process.append(
+                    parallel_state.is_pipeline_last_stage(ignore_virtual=False, vp_stage=vp_stage)
+                )
+        else:
+            model_post_process.append(parallel_state.is_pipeline_last_stage())
+
+        if len(model_post_process) != len(model_chunks):
+            raise ValueError(
+                "Model list length and pipeline post-process list length must match. "
+                f"Got {len(model_chunks)} model chunks and {len(model_post_process)} post-process flags."
+            )
+
+        for index, model_chunk in enumerate(model_chunks):
+            if model_post_process[index]:
+                model_chunk.output_layer = LinearForLastLayer(
+                    input_size=hidden_size,
+                    output_size=output_size,
+                    sequence_parallel=sequence_parallel,
+                )
+
+        return model_chunks
+
+    return hook
+
+
+def make_value_model(hidden_size: int, sequence_parallel: bool) -> ModelHook:
+    """Create a value-head hook compatible with existing external trainer code."""
+    return create_value_head_hook(hidden_size=hidden_size, sequence_parallel=sequence_parallel)
+
+
+def freeze_moe_router(model: ModelList | MegatronModule) -> ModelList:
+    """Freeze MoE router and shared-expert gate parameters in model chunks.
+
+    Args:
+        model: Single Megatron module or list of virtual-pipeline model chunks.
+
+    Returns:
+        The normalized model chunk list with router parameters frozen in place.
+    """
+    model_chunks = _ensure_model_list(model)
+    for model_chunk in model_chunks:
+        decoder = getattr(model_chunk, "decoder", None)
+        layers = getattr(decoder, "layers", None)
+        if layers is None:
+            continue
+        for layer in layers:
+            mlp = getattr(layer, "mlp", None)
+            if mlp is None:
+                continue
+            router = getattr(mlp, "router", None)
+            if router is not None:
+                _freeze_parameter_if_present(router, "weight")
+                _freeze_parameter_if_present(router, "bias")
+
+            shared_experts = getattr(mlp, "shared_experts", None)
+            if shared_experts is not None:
+                _freeze_parameter_if_present(shared_experts, "gate_weight")
+                _freeze_parameter_if_present(shared_experts, "gate_bias")
+
+    return model_chunks
+
+
+def _ensure_model_list(model: ModelList | MegatronModule) -> ModelList:
+    return model if isinstance(model, list) else [model]
+
+
+def _freeze_parameter_if_present(module: object, name: str) -> None:
+    parameter = getattr(module, name, None)
+    if parameter is not None:
+        parameter.requires_grad = False
+
+
+def _register_linear_for_last_layer_mapping() -> None:
+    from megatron.bridge.models.conversion.param_mapping import AutoMapping
+
+    AutoMapping.register_module_type("LinearForLastLayer", "replicated")
 
 
 def start_memory_history_recording(profiling: ProfilingConfig | None) -> None:
